pi4j version to 2.3.0

+++ grovepi-spec/src/main/java/com/github/yafna/raspberry/grovepi/devices/GroveUltrasonicRanger.java
@@ -19,7 +19,7 @@ public class GroveUltrasonicRanger {
   }
 
   public double get() throws IOException {
-    return grovePi.exec((io) -> {
+    return (double) grovePi.exec((io) -> {
       io.write(GrovePiCommands.uRead_cmd, pin, GrovePiCommands.unused, GrovePiCommands.unused);
       io.sleep(200);
       int[] v = GroveUtil.unsign(io.read(new byte[4]));
